pyproject.toml
[tool.poetry]
name = "projeto portal transparência"
version = "0.1.0"
description = "projeto para analise de dados do portal da transparência"
authors = ["André Mendes Nunes <branco.de@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.2.2"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
